add own action for icon deletion
