--- app/Filament/Components/Actions/DeleteServerIcon.php
@@ -1,48 +1,0 @@
-<?php
-
-namespace App\Filament\Components\Actions;
-
-use App\Enums\TablerIcon;
-use App\Models\Server;
-use Filament\Actions\Action;
-use Filament\Notifications\Notification;
-use Illuminate\Support\Facades\Storage;
-
-class DeleteServerIcon extends Action
-{
-    public static function getDefaultName(): ?string
-    {
-        return 'delete_icon';
-    }
-
-    protected function setUp(): void
-    {
-        parent::setUp();
-
-        $this->visible(fn ($record) => $record->icon);
-
-        $this->hiddenLabel();
-
-        $this->tooltip(trans('admin/server.import_icon'));
-
-        $this->icon(TablerIcon::Trash);
-
-        $this->color('danger');
-
-        $this->action(function ($record) {
-            foreach (array_keys(Server::ICON_FORMATS) as $ext) {
-                $path = Server::ICON_STORAGE_PATH . "/$record->uuid.$ext";
-                if (Storage::disk('public')->exists($path)) {
-                    Storage::disk('public')->delete($path);
-                }
-            }
-
-            Notification::make()
-                ->title(trans('server/setting.server_info.icon.deleted'))
-                ->success()
-                ->send();
-
-            $record->refresh();
-        });
-    }
-}
